Enable code completion based on trait bounds for function args
It's rebump of https://github.com/racer-rust/racer/pull/828.
As a first aid to libsyntax's change, I turned off this feature, but
now modified codes in accordance with libsyntax. Those changes are
very straight forward.
In addition, this commit includes removal of unused codes and some
refactoring of related codes.